extracted some code from errorlogger.h to errortypes.h
